[Update] - added Relay peer to project
