--- go-code/peer_manager.go
@@ -1,8 +1,0 @@
-package main
-
-// PeerManager struct to manage peer state
-type PeerManager struct {
-	done        chan bool // Channel to signal termination
-	disconnect  chan bool // Channel to signal termination
-	unsubscribe chan bool // Channel to signal termination
-}
